We add a new API Key method and views on front end
This is a partial edit, here we move in the direction of making API Keys
a full and proper model, and include them properly. The necessary views
for it are created, and they are displayed properly. Also the routes for
index, create and destroy are created.

However front end functions for creation and deletion have not been
yet written and hence the buttons do nothing.

Major change was taking API Keys out of Me View and adding it as a composite
view