Add a front end method for adding API Keys
This adds a front end method which makes a POST call to add an API Key
to a particular user path. This has not been done in the backend, just
the front end. So it results in a 404.

This basically adds a empty container and if successful keeps it, else
pops it