Add the backend method for adding API Keys
This adds the corresponding backend methods for creating API keys,
this also changed the front end to behave properly. The behaviour now is
* User clicks button
* Button is disabled and loading
* The view is refreshed once it is finished loading

We moved the model add to collection from after creating to after save
which in hindsight is logical. We also had to make a few typo changes