Add a frontend method for deleting API Keys
This deletes API Keys from the user, and has a Delete? Yes, No
Confirmation secion. Once confirmed, a destroy action is called.
The backend for this is yet to be implemented, and hence returns 500

The major unexpected roadblock here was that id was not getting properly
set because of moving to MongoDB, so we wrote a as_json to override it
and got it working.

The backend needs to be done