Add a new List lectures method instead of the old one
We have moved to lectures being embedded inside courses, and as a result
we are also changing the lectures method.

We also moved the list methods to lectures module from Courses module
for the simple reason that it was logically suited there, and similar
to courses/new we use App.execute to call it.

This is actually incomplete we will be getting back to this once we fix
down on lecture new method.